Server Side (API +Image Processing Module)/requirements.txt
asgiref==3.2.7
cycler==0.10.0
Cython==0.29.17
Django==3.0.6
djangorestframework==3.11.0
imutils==0.5.3
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
tesseract==0.1.3
